Split index code into multiple files
